% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{adjusted_r_squared}
\alias{adjusted_r_squared}
\title{Adjusted coefficient of determination}
\usage{
adjusted_r_squared(y_true, y_pred, d)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}

\item{d}{Number of predictors in the downstream model.}
}
\value{
Numeric adjusted R-squared.
}
\description{
Adjusted coefficient of determination
}
\examples{
adjusted_r_squared(
  c(2, 4, 6, 8, 10),
  c(2.2, 3.8, 5.7, 8.1, 9.8),
  d = 1
)
}
