% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{ar1_covariance}
\alias{ar1_covariance}
\title{AR(1) covariance matrix}
\usage{
ar1_covariance(p, rho)
}
\arguments{
\item{p}{Dimension.}

\item{rho}{Correlation parameter.}
}
\value{
Covariance matrix.
}
\description{
AR(1) covariance matrix
}
\examples{
ar1_covariance(p = 4, rho = 0.5)
}
