% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_stabilisation.R
\name{check_cov_matrix}
\alias{check_cov_matrix}
\title{Check covariance matrix}
\usage{
check_cov_matrix(Sigma)
}
\arguments{
\item{Sigma}{Matrix.}
}
\value{
A numeric square symmetric matrix.
}
\description{
Internal function for validating covariance matrix input.
}
\keyword{internal}
