% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_cov_method}
\alias{check_cov_method}
\title{Check covariance method}
\usage{
check_cov_method(method)
}
\arguments{
\item{method}{Character string.}
}
\value{
Matched covariance method.
}
\description{
Check covariance method
}
\keyword{internal}
