% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_flag}
\alias{check_flag}
\title{Validate a logical flag}
\usage{
check_flag(x, name)
}
\arguments{
\item{x}{Object to validate.}

\item{name}{Argument name used in error messages.}
}
\value{
The validated logical value.
}
\description{
Validate a logical flag
}
\keyword{internal}
