% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_missing}
\alias{check_missing}
\title{Check missing values}
\usage{
check_missing(X, y)
}
\arguments{
\item{X}{Predictor matrix.}

\item{y}{Response vector.}
}
\value{
Invisibly returns TRUE if no missing values are found.
}
\description{
Internal function for detecting missing values.
}
\keyword{internal}
