% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_named_list}
\alias{check_named_list}
\title{Validate a named argument list}
\usage{
check_named_list(x, name)
}
\arguments{
\item{x}{Object to validate.}

\item{name}{Argument name used in error messages.}
}
\value{
The validated list.
}
\description{
Validate a named argument list
}
\keyword{internal}
