% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_nslices}
\alias{check_nslices}
\title{Check number of slices}
\usage{
check_nslices(nslices, n)
}
\arguments{
\item{nslices}{Number of response slices.}

\item{n}{Sample size.}
}
\value{
Integer number of slices.
}
\description{
Check number of slices
}
\keyword{internal}
