% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{check_prediction_vectors}
\alias{check_prediction_vectors}
\title{Check prediction vectors}
\usage{
check_prediction_vectors(y_true, y_pred)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}
}
\value{
Invisibly TRUE.
}
\description{
Check prediction vectors
}
\keyword{internal}
