% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_sdr_method}
\alias{check_sdr_method}
\title{Check SDR method}
\usage{
check_sdr_method(method)
}
\arguments{
\item{method}{Character string.}
}
\value{
Matched SDR method.
}
\description{
Check SDR method
}
\keyword{internal}
