% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_selector}
\alias{check_selector}
\title{Check selector}
\usage{
check_selector(selector)
}
\arguments{
\item{selector}{Model selection criterion.}
}
\value{
Matched selector.
}
\description{
Check selector
}
\keyword{internal}
