% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_stabilization_method}
\alias{check_stabilization_method}
\title{Check stabilisation method}
\usage{
check_stabilization_method(method)
}
\arguments{
\item{method}{Character string.}
}
\value{
Matched stabilisation method.
}
\description{
Check stabilisation method
}
\keyword{internal}
