% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_X}
\alias{check_X}
\title{Check predictor matrix}
\usage{
check_X(X, min_rows = 5L, min_cols = 2L)
}
\arguments{
\item{X}{A numeric matrix or data frame of predictors.}

\item{min_rows}{Minimum permitted number of rows.}

\item{min_cols}{Minimum permitted number of columns.}
}
\value{
A numeric matrix.
}
\description{
Internal function for validating predictor input.
}
\keyword{internal}
