% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{check_y_continuous}
\alias{check_y_continuous}
\title{Check continuous response vector}
\usage{
check_y_continuous(y, n)
}
\arguments{
\item{y}{Numeric response vector.}

\item{n}{Expected sample size.}
}
\value{
Numeric response vector.
}
\description{
Internal function for validating continuous response input.
}
\keyword{internal}
