% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_selection.R
\name{choose_dimension}
\alias{choose_dimension}
\title{Choose dimension from criteria table}
\usage{
choose_dimension(
  d_table,
  selector = c("cicomp", "icomp", "bic", "caic", "aic")
)
}
\arguments{
\item{d_table}{Data frame returned by select_dimension().}

\item{selector}{Criterion used for selection.}
}
\value{
Selected structural dimension.
}
\description{
Choose dimension from criteria table
}
\examples{
scores <- as.matrix(mtcars[, c("wt", "hp", "disp")])
dimension_table <- select_dimension(scores, mtcars$mpg, d_max = 3)
choose_dimension(dimension_table, selector = "bic")
}
