% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_kernels.R
\name{compute_phd}
\alias{compute_phd}
\title{Compute Principal Hessian Directions}
\usage{
compute_phd(X, y, Sigma = NULL, eps = 1e-08)
}
\arguments{
\item{X}{Numeric predictor matrix.}

\item{y}{Numeric response vector.}

\item{Sigma}{Covariance matrix used for standardisation.}

\item{eps}{Eigenvalue floor.}
}
\value{
A list containing pHd kernel, eigenvalues, directions, scores.
}
\description{
Compute Principal Hessian Directions
}
\examples{
X <- as.matrix(mtcars[, c("disp", "hp", "wt")])
fit <- compute_phd(X, y = mtcars$mpg)
head(fit$eigenvalues)
}
