% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{compute_scores}
\alias{compute_scores}
\title{Compute SDR scores for new data}
\usage{
compute_scores(newX, directions)
}
\arguments{
\item{newX}{New predictor matrix or data frame.}

\item{directions}{Matrix of SDR directions.}
}
\value{
Matrix of SDR scores.
}
\description{
Projects new predictor observations onto estimated SDR directions.
}
