% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_kernels.R
\name{compute_sdr}
\alias{compute_sdr}
\title{General SDR kernel dispatcher}
\usage{
compute_sdr(
  X,
  y,
  method = c("dr", "sir", "save", "phd"),
  Sigma = NULL,
  nslices = 6,
  ...
)
}
\arguments{
\item{X}{Numeric predictor matrix.}

\item{y}{Numeric response vector.}

\item{method}{SDR method.}

\item{Sigma}{Covariance matrix.}

\item{nslices}{Number of slices.}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
SDR fit components.
}
\description{
General SDR kernel dispatcher
}
