% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_stabilisation.R
\name{cov_effective_rank}
\alias{cov_effective_rank}
\title{Effective covariance rank}
\usage{
cov_effective_rank(Sigma, tol = 1e-06)
}
\arguments{
\item{Sigma}{A covariance matrix.}

\item{tol}{Positive threshold.}
}
\value{
Effective rank.
}
\description{
Computes the number of eigenvalues exceeding a threshold.
}
\examples{
Sigma <- cov(mtcars[, c("disp", "hp", "wt")])
cov_effective_rank(Sigma)
}
