% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_estimators.R
\name{cov_sample}
\alias{cov_sample}
\title{Sample covariance matrix}
\usage{
cov_sample(X)
}
\arguments{
\item{X}{Numeric matrix or data frame.}
}
\value{
A covariance matrix.
}
\description{
Computes the usual unbiased sample covariance matrix.
}
