% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_selection.R
\name{criterion_weights}
\alias{criterion_weights}
\title{Information criterion weights}
\usage{
criterion_weights(
  d_table,
  criterion = c("AIC", "BIC", "CAIC", "ICOMP", "CICOMP")
)
}
\arguments{
\item{d_table}{Dimension selection table.}

\item{criterion}{Criterion column name.}
}
\value{
Data frame with criterion differences and weights.
}
\description{
Computes relative weights from any information criterion column.
}
