% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_kernels.R
\name{decompose_kernel}
\alias{decompose_kernel}
\title{Eigen-decompose SDR kernel}
\usage{
decompose_kernel(M, sort_by_abs = FALSE)
}
\arguments{
\item{M}{Kernel matrix.}

\item{sort_by_abs}{Logical. If TRUE, sorts by absolute eigenvalue magnitude.}
}
\value{
A list with kernel, eigenvalues, directions.
}
\description{
Eigen-decompose SDR kernel
}
\keyword{internal}
