% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{evaluate_prediction}
\alias{evaluate_prediction}
\title{Evaluate predictions}
\usage{
evaluate_prediction(y_true, y_pred, d = NULL)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}

\item{d}{Optional number of reduced predictors.}
}
\value{
A data frame of prediction metrics.
}
\description{
Computes common prediction metrics.
}
