% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_risdr.R
\name{extract_loadings}
\alias{extract_loadings}
\title{Extract SDR loadings}
\usage{
extract_loadings(directions, variables = NULL)
}
\arguments{
\item{directions}{Matrix of SDR directions.}

\item{variables}{Variable names.}
}
\value{
A data frame of loadings.
}
\description{
Extract SDR loadings
}
