% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata_workflow.R
\name{filter_low_variance}
\alias{filter_low_variance}
\title{Filter low-variance predictors}
\usage{
filter_low_variance(X, variance_quantile = 0.25)
}
\arguments{
\item{X}{Predictor matrix.}

\item{variance_quantile}{Quantile threshold.}
}
\value{
Filtered matrix.
}
\description{
Removes predictors with variance below a specified quantile.
}
