% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{fit_downstream_lm}
\alias{fit_downstream_lm}
\title{Fit downstream regression model on SDR scores}
\usage{
fit_downstream_lm(scores, y, d)
}
\arguments{
\item{scores}{Matrix of SDR scores.}

\item{y}{Numeric response vector.}

\item{d}{Structural dimension.}
}
\value{
A fitted lm object.
}
\description{
Fits a linear regression model using the first d SDR scores.
}
\examples{
scores <- as.matrix(mtcars[, c("wt", "hp")])
fit_downstream_lm(scores, y = mtcars$mpg, d = 2)
}
