% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_signal}
\alias{generate_signal}
\title{Generate SDR signal}
\usage{
generate_signal(
  U,
  model = c("linear_quadratic", "symmetric", "interaction", "linear")
)
}
\arguments{
\item{U}{Matrix of sufficient predictors.}

\item{model}{Signal model.}
}
\value{
Numeric signal vector.
}
\description{
Generate SDR signal
}
\keyword{internal}
