% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(y_true, y_pred)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}
}
\value{
Numeric MAE.
}
\description{
Mean absolute error
}
