% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{make_cv_folds}
\alias{make_cv_folds}
\title{Create cross-validation folds}
\usage{
make_cv_folds(n, v = 5, seed = NULL)
}
\arguments{
\item{n}{Sample size.}

\item{v}{Number of folds.}

\item{seed}{Optional random seed.}
}
\value{
Integer vector of fold memberships.
}
\description{
Create cross-validation folds
}
