% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{make_true_beta}
\alias{make_true_beta}
\title{Construct true central subspace basis}
\usage{
make_true_beta(
  p,
  d,
  beta_type = c("coordinate", "random_sparse", "random_dense")
)
}
\arguments{
\item{p}{Number of predictors.}

\item{d}{Structural dimension.}

\item{beta_type}{Basis type.}
}
\value{
A p x d orthonormal basis matrix.
}
\description{
Construct true central subspace basis
}
\keyword{internal}
