% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{mape}
\alias{mape}
\title{Mean absolute percentage error}
\usage{
mape(y_true, y_pred)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}
}
\value{
Numeric MAPE.
}
\description{
Mean absolute percentage error
}
\examples{
mape(c(2, 4, 6, 8), c(2.2, 3.8, 5.7, 8.1))
}
