% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdr_kernels.R
\name{matrix_inv_sqrt}
\alias{matrix_inv_sqrt}
\title{Matrix inverse square root}
\usage{
matrix_inv_sqrt(Sigma, eps = 1e-08)
}
\arguments{
\item{Sigma}{Symmetric positive definite matrix.}

\item{eps}{Eigenvalue floor.}
}
\value{
Matrix inverse square root.
}
\description{
Computes the inverse square root of a symmetric positive definite matrix.
}
\keyword{internal}
