% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{plot_cv_dimension}
\alias{plot_cv_dimension}
\title{Plot cross-validation dimension selection result}
\usage{
plot_cv_dimension(cv, metric = NULL, ...)
}
\arguments{
\item{cv}{Object returned by select_dimension_cv().}

\item{metric}{Metric to plot.}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
Invisibly returns the CV table.
}
\description{
Plot cross-validation dimension selection result
}
