% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dimension_selection}
\alias{plot_dimension_selection}
\title{Plot dimension selection criteria}
\usage{
plot_dimension_selection(
  fit,
  criteria = c("AIC", "BIC", "CAIC", "ICOMP", "CICOMP"),
  ...
)
}
\arguments{
\item{fit}{Object of class "risdr".}

\item{criteria}{Character vector of criteria to plot.}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:matplot]{graphics::matplot()}}.}
}
\value{
Invisibly returns dimension table.
}
\description{
Plot dimension selection criteria
}
