% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ladle.R
\name{plot_ladle}
\alias{plot_ladle}
\title{Plot ladle dimension diagnostic}
\usage{
plot_ladle(ladle, ...)
}
\arguments{
\item{ladle}{Object returned by select_dimension_ladle().}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:matplot]{graphics::matplot()}}.}
}
\value{
Invisibly returns ladle table.
}
\description{
Plot ladle dimension diagnostic
}
