% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Plot SDR loadings}
\usage{
plot_loadings(fit, direction = 1, top = 15, ...)
}
\arguments{
\item{fit}{Object of class "risdr".}

\item{direction}{Direction index.}

\item{top}{Number of variables to display.}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:barplot]{graphics::barplot()}}.}
}
\value{
Invisibly returns loading table.
}
\description{
Displays the largest absolute loadings for a selected SDR direction.
}
