% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_prediction}
\alias{plot_prediction}
\title{Prediction plot}
\usage{
plot_prediction(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns plotted data.
}
\description{
Plots observed versus predicted values.
}
