% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Residual plot}
\usage{
plot_residuals(y_true, y_pred, ...)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns plotted data.
}
\description{
Plots residuals against predicted values.
}
