% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scree}
\alias{plot_scree}
\title{Scree plot of SDR eigenvalues}
\usage{
plot_scree(fit, n_eigen = 20, type = "b", ...)
}
\arguments{
\item{fit}{Object of class "risdr".}

\item{n_eigen}{Number of eigenvalues to display.}

\item{type}{Plot type passed to base R plot().}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns eigenvalues plotted.
}
\description{
Scree plot of SDR eigenvalues
}
