% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sufficient}
\alias{plot_sufficient}
\title{Sufficient summary plot}
\usage{
plot_sufficient(fit, direction = 1, ...)
}
\arguments{
\item{fit}{Object of class "risdr".}

\item{direction}{SDR direction to plot.}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns plotted data.
}
\description{
Plots the response against a selected SDR direction.
}
