% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_sufficient2d}
\alias{plot_sufficient2d}
\title{Two-direction sufficient summary plot}
\usage{
plot_sufficient2d(
  fit,
  directions = c(1, 2),
  colour_by_y = TRUE,
  groups = 4,
  ...
)
}
\arguments{
\item{fit}{Object of class "risdr".}

\item{directions}{Integer vector of length 2.}

\item{colour_by_y}{Logical. If TRUE, colours points by response quantile group.}

\item{groups}{Number of response colour groups.}

\item{...}{Additional graphical arguments passed to \code{\link[graphics:plot.default]{graphics::plot()}}.}
}
\value{
Invisibly returns plotted data.
}
\description{
Plots the first two selected SDR scores, optionally coloured by response.
}
