% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_downstream_lm}
\alias{predict_downstream_lm}
\title{Predict from downstream SDR regression model}
\usage{
predict_downstream_lm(fit_lm, scores_new, d)
}
\arguments{
\item{fit_lm}{Fitted linear model from fit_downstream_lm().}

\item{scores_new}{Matrix of new SDR scores.}

\item{d}{Structural dimension.}
}
\value{
Numeric vector of predictions.
}
\description{
Predict from downstream SDR regression model
}
