% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction_correlation}
\alias{prediction_correlation}
\title{Prediction correlation}
\usage{
prediction_correlation(y_true, y_pred)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}
}
\value{
Pearson correlation.
}
\description{
Prediction correlation
}
