% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata_workflow.R
\name{prepare_survival_response}
\alias{prepare_survival_response}
\title{Prepare survival response}
\usage{
prepare_survival_response(time, delta)
}
\arguments{
\item{time}{Survival time vector.}

\item{delta}{Event indicator vector.}
}
\value{
Cleaned survival response list.
}
\description{
Prepare survival response
}
