% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_risdr.R
\name{print.risdr}
\alias{print.risdr}
\title{Print risdr object}
\usage{
\method{print}{risdr}(x, ...)
}
\arguments{
\item{x}{Object of class "risdr".}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
Invisibly returns x.
}
\description{
Print risdr object
}
