% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_engine.R
\name{print.risdr_dual}
\alias{print.risdr_dual}
\title{Print method for dual RISDR objects}
\usage{
\method{print}{risdr_dual}(x, ...)
}
\arguments{
\item{x}{Object of class \code{risdr_dual}.}

\item{...}{Additional arguments passed to internal methods.}
}
\description{
Print method for dual RISDR objects
}
