% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realdata_workflow.R
\name{print.risdr_realdata}
\alias{print.risdr_realdata}
\title{Print real-data RISDR workflow}
\usage{
\method{print}{risdr_realdata}(x, ...)
}
\arguments{
\item{x}{Object of class risdr_realdata.}

\item{...}{Additional arguments.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print real-data RISDR workflow
}
