% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_engine.R
\name{print.risdr_sparse_realdata}
\alias{print.risdr_sparse_realdata}
\title{Print real-data RISDR workflow}
\usage{
\method{print}{risdr_sparse_realdata}(x, ...)
}
\arguments{
\item{x}{Object of class risdr_realdata.}

\item{...}{Additional arguments.}
}
\description{
Print real-data RISDR workflow
}
