% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_risdr.R
\name{print.summary.risdr}
\alias{print.summary.risdr}
\title{Print summary of risdr object}
\usage{
\method{print}{summary.risdr}(x, ...)
}
\arguments{
\item{x}{Object of class "summary.risdr".}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
Invisibly returns x.
}
\description{
Print summary of risdr object
}
