% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_engine.R
\name{print.summary.risdr_sparse}
\alias{print.summary.risdr_sparse}
\title{Print summary of risdr object}
\usage{
\method{print}{summary.risdr_sparse}(x, ...)
}
\arguments{
\item{x}{Object of class "summary.risdr_sparse".}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
Invisibly returns x.
}
\description{
Print summary of risdr object
}
