% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{rownames_or_names}
\alias{rownames_or_names}
\title{Helper for row names or fallback names}
\usage{
rownames_or_names(x, fallback = NULL)
}
\arguments{
\item{x}{Matrix.}

\item{fallback}{Fallback names.}
}
\value{
Character vector.
}
\description{
Helper for row names or fallback names
}
\keyword{internal}
