% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_mvn}
\alias{simulate_mvn}
\title{Simulate multivariate normal data}
\usage{
simulate_mvn(n, Sigma)
}
\arguments{
\item{n}{Sample size.}

\item{Sigma}{Covariance matrix.}
}
\value{
Numeric matrix.
}
\description{
Simulate multivariate normal data
}
\keyword{internal}
