% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{slice_means}
\alias{slice_means}
\title{Compute slice means}
\usage{
slice_means(X, slices)
}
\arguments{
\item{X}{Numeric matrix.}

\item{slices}{Integer slice memberships.}
}
\value{
Matrix of slice means with rows corresponding to slices.
}
\description{
Computes slice-specific means of predictors.
}
