% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{slice_proportions}
\alias{slice_proportions}
\title{Compute slice proportions}
\usage{
slice_proportions(slices)
}
\arguments{
\item{slices}{Integer slice memberships.}
}
\value{
Numeric vector of slice proportions.
}
\description{
Computes empirical slice probabilities.
}
\examples{
slices <- make_slices(mtcars$mpg, nslices = 4)
slice_proportions(slices)
}
