% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_stabilisation.R
\name{stabilize_eigenfloor}
\alias{stabilize_eigenfloor}
\title{Eigenvalue floor stabilisation}
\usage{
stabilize_eigenfloor(Sigma, eps = 1e-06)
}
\arguments{
\item{Sigma}{A square symmetric covariance matrix.}

\item{eps}{Positive eigenvalue floor.}
}
\value{
A symmetric positive definite covariance matrix.
}
\description{
Stabilises a symmetric covariance matrix by replacing eigenvalues smaller than
a positive threshold with that threshold.
}
