% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_stabilisation.R
\name{stabilize_ridge}
\alias{stabilize_ridge}
\title{Ridge stabilisation of covariance matrix}
\usage{
stabilize_ridge(Sigma, lambda = 1e-04)
}
\arguments{
\item{Sigma}{A square symmetric covariance matrix.}

\item{lambda}{Ridge stabilisation parameter.}
}
\value{
A symmetric positive definite covariance matrix.
}
\description{
Stabilises a covariance matrix by adding a positive multiple of the identity matrix.
}
\examples{
Sigma <- matrix(c(1, 1, 1, 1), nrow = 2)
eigen(stabilize_ridge(Sigma))$values
}
