% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs.R
\name{standardize_X}
\alias{standardize_X}
\title{Standardise predictors}
\usage{
standardize_X(X, center = NULL, scale = NULL)
}
\arguments{
\item{X}{Numeric predictor matrix.}

\item{center}{Optional centring vector.}

\item{scale}{Optional scaling vector.}
}
\value{
A list containing standardised matrix, centre, and scale.
}
\description{
Standardise predictors
}
\keyword{internal}
