% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{subspace_distance}
\alias{subspace_distance}
\title{Subspace distance}
\usage{
subspace_distance(B_hat, B)
}
\arguments{
\item{B_hat}{Estimated basis matrix.}

\item{B}{True basis matrix.}
}
\value{
Numeric subspace distance.
}
\description{
Computes Frobenius distance between projection matrices.
}
\examples{
B <- matrix(c(1, 0, 0, 0, 1, 0), nrow = 3)
transformed <- B \%*\% diag(c(2, 3))
subspace_distance(transformed, B)
}
