% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{summarise_simulation}
\alias{summarise_simulation}
\title{Summarise simulation results}
\usage{
summarise_simulation(sim_results)
}
\arguments{
\item{sim_results}{Data frame from run_risdr_simulation().}
}
\value{
Summary data frame.
}
\description{
Summarise simulation results
}
