% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_risdr.R
\name{summary.risdr}
\alias{summary.risdr}
\title{Summarise risdr object}
\usage{
\method{summary}{risdr}(object, ...)
}
\arguments{
\item{object}{Object of class "risdr".}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
A list summary.
}
\description{
Summarise risdr object
}
