% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_engine.R
\name{summary.risdr_dual}
\alias{summary.risdr_dual}
\title{Summary method for dual RISDR objects}
\usage{
\method{summary}{risdr_dual}(object, top_n = 20L, ...)
}
\arguments{
\item{object}{Object of class \code{risdr_dual}.}

\item{top_n}{Number of selected variables to print.}

\item{...}{Additional arguments passed to internal methods.}
}
\description{
Summary method for dual RISDR objects
}
