% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis_engine.R
\name{summary.risdr_sparse}
\alias{summary.risdr_sparse}
\title{Summarise risdr object}
\usage{
\method{summary}{risdr_sparse}(object, ...)
}
\arguments{
\item{object}{Object of class "risdr".}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
A list summary.
}
\description{
Summarise risdr object
}
