% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_risdr.R
\name{predict.risdr}
\alias{predict.risdr}
\title{Predict method for risdr objects}
\usage{
\method{predict}{risdr}(object, newX, d = NULL, ...)
}
\arguments{
\item{object}{Object of class "risdr".}

\item{newX}{New predictor matrix or data frame.}

\item{d}{Optional structural dimension.}

\item{...}{Additional arguments passed to internal methods.}
}
\value{
Numeric vector of predictions.
}
\description{
Predict method for risdr objects
}
