% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{prediction_correlation}
\alias{prediction_correlation}
\title{Prediction correlation}
\usage{
prediction_correlation(y_true, y_pred)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}
}
\value{
Pearson correlation.
}
\description{
Prediction correlation
}
\examples{
prediction_correlation(
  c(2, 4, 6, 8),
  c(2.2, 3.8, 5.7, 8.1)
)
}
