% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{projection_matrix}
\alias{projection_matrix}
\title{Projection matrix}
\usage{
projection_matrix(B)
}
\arguments{
\item{B}{Basis matrix.}
}
\value{
Projection matrix.
}
\description{
Projection matrix
}
\examples{
B <- matrix(c(1, 0, 0, 0, 1, 0), nrow = 3)
projection_matrix(B)
}
