% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{r_squared}
\alias{r_squared}
\title{Coefficient of determination}
\usage{
r_squared(y_true, y_pred)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}
}
\value{
Numeric R-squared.
}
\description{
Coefficient of determination
}
