% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(y_true, y_pred)
}
\arguments{
\item{y_true}{Observed response values.}

\item{y_pred}{Predicted response values.}
}
\value{
Numeric RMSE.
}
\description{
Root mean squared error
}
